refactor file; add graph tool;

+++ UtilityLib/VecUtil.h
@@ -113,7 +113,7 @@ template <typename Iter_1,typename Iter_2>
 inline void SubVec( Iter_1 vec_begin, Iter_2 idx_begin, Iter_1 result_begin, size_t n)
 {
 	static_assert( std::_Is_random_iter_v<Iter_1> || std::is_array_v<Iter_1> );
-	for( int i = 0; i < n; i++ )
+	for( int i = 0; i < (int)n; i++ )
 	{
 		int idx = *idx_begin++;
 		*result_begin++ = *( vec_begin + idx );
